LayoutTests/storage/indexeddb/transaction-and-objectstore-calls-expected.txt

Test IndexedDB's transaction and objectStore calls

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'webkitIndexedDB' in window is true
PASS webkitIndexedDB == null is false
webkitIndexedDB.open('name', 'description')
PASS 'onsuccess' in result is true
PASS 'onerror' in result is true
PASS 'readyState' in result is true
An event should fire shortly...

db = event.result
result = db.setVersion('version 1')
trans = event.result
Deleted all object stores.
db.createObjectStore('a')
db.createObjectStore('b')
trans.oncomplete = created

trans = db.transaction({objectStoreNames: 'a'})
trans.objectStore('a')
Expecting exception from trans.objectStore('b')
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR
Expecting exception from trans.objectStore('x')
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR

trans = db.transaction({objectStoreNames: ['a']})
trans.objectStore('a')
Expecting exception from trans.objectStore('b')
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR
Expecting exception from trans.objectStore('x')
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR

trans = db.transaction({objectStoreNames: ['b']})
trans.objectStore('b')
Expecting exception from trans.objectStore('a')
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR
Expecting exception from trans.objectStore('x')
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR

trans = db.transaction({objectStoreNames: ['a', 'b']})
trans.objectStore('a')
trans.objectStore('b')
Expecting exception from trans.objectStore('x')
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR

trans = db.transaction({objectStoreNames: ['b', 'a']})
trans.objectStore('a')
trans.objectStore('b')
Expecting exception from trans.objectStore('x')
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR

trans = db.transaction({objectStoreNames: []})
trans.objectStore('a')
trans.objectStore('b')
Expecting exception from trans.objectStore('x')
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR

trans = db.transaction()
trans.objectStore('a')
trans.objectStore('b')
Expecting exception from trans.objectStore('x')
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR

Expecting exception from db.transaction({objectStoreNames: 'x'})
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR
Expecting exception from db.transaction({objectStoreNames: ['x']})
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR
Expecting exception from db.transaction({objectStoreNames: ['a', 'x']})
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR
Expecting exception from db.transaction({objectStoreNames: ['x', 'x']})
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR
Expecting exception from db.transaction({objectStoreNames: ['a', 'x', 'b']})
PASS code is webkitIDBDatabaseException.NOT_FOUND_ERR

PASS successfullyParsed is true

TEST COMPLETE
